Moved and consolidated some components
