+++ src/app/pastTrips/layout.tsx
@@ -1,6 +1,6 @@
 'use server';
 import { getUser } from '@workos-inc/authkit-nextjs';
-import TripContainer from '@/app/containers/trips/tripContainer';
+import TripContainer from '@/components/trips/tripContainer';
 
 export default async function CurrentTrip({
   children,
