+++ src/components/trips/basicDisplayTrips.tsx
