+++ src/components/trips/inDepthDisplayTrip.tsx
