+++ src/components/trips/tripContainer.tsx
